==> 02_sampling_labeling.ipynb <==
# Sampling Gmail Data

Although we have around 70k mails, we will need 1000-1500 well-labeled mails for the training purposes.
> **Label Quality matters far more than Volume.** 

## Objective

The sampling cannot be done randomly because we'll get :

- Mostly similar promotional and social media emails (more than 80-90% of the mails are promotional and social)
- Very few truly important emails (quantity is less)
- Almost no obvious spam (the quantity is already less than 100)

That's **bad training data** because the ML Model 

- Barely sees important emails
- Almost no spam examples
- Classifier becomes biased

> Random Sampling reflects reality but fails to teach the model how to to make predictions.



We'll adopt an intelligent Sampling Strategy, focusing on diversity across :

- Gmail label
- Sender Domain
- Email intent
- Time

The goal is to

> **Expose the model to different “types” of emails you receive.**

## Stratified Sampling

In our case, since 
- the gmail spam folder is tiny
- promotions and social media mails dominate
- important mails are fewer but critical

We'll adopt a strategy called Stratified Sampling.

> **Stratified sampling is a way of sampling data so that important subgroups (strata) are properly represented in your sample.**


1. Split the data into meaningful groups
2. Take random samples from each group

Instead of picking points purely at random, you **control the composition of the sample.**


## Exploring the data

#Loading the emails data into a Pandas dataframe
import pandas as pd

df = pd.read_csv("../data/raw_emails.csv")
print("Total emails:", len(df))
df.head()


df["labels"].value_counts()

## These are the different categories and labels. Almost 95% of my emails are unread. Clearly,I don't really read my emails.
## And almost 50% of them are social media updates.We have apps for social media updates, why are they sending me mails.

df["has_attachment"].value_counts()

## None of the emails have any attachments.

df["sender_domain"].value_counts()

## Facebook sends me a lot of emails. I am not even active on it. Maybe that's why.

## Understanding Gmail Labels

In Gmail, everything is a label.

- Inbox, Spam, Promotions → all are labels
- Emails can have multiple labels at once
- Categories are just special system labels

So technically:

>“Categories are labels with special meaning.”

#### Core system labels (used internally by Gmail)

These are always present and very reliable.

| Label       | Meaning                        |
| ----------- | ------------------------------ |
| `INBOX`     | Appears in inbox               |
| `SPAM`      | Classified as spam             |
| `TRASH`     | Deleted                        |
| `UNREAD`    | Not opened                     |
| `STARRED`   | Starred by user                |
| `IMPORTANT` | Gmail thinks it matters to you |
| `SENT`      | Sent by you                    |
| `DRAFT`     | Draft email                    |

#### Category labels (the tabs you see in Gmail)

These are content-based classifications.

| Category   | Label ID              | Meaning | Relevance
| ---------- | --------------------- |----------| ----------|
| Primary    | `CATEGORY_PERSONAL`   |Person-to-person emails, Bank alerts, OTPs, Work / important communication| *High likelihood you’ll open it*|
| Promotions | `CATEGORY_PROMOTIONS` |Universities, Marketing emails, Discounts, newsletters, Brand communication| *Legit, but not urgent*|
| Social     | `CATEGORY_SOCIAL`     |LinkedIn, Facebook, Instagram, Twitter notifications|*Engagement-driven*|
| Updates    | `CATEGORY_UPDATES`    |Bills, Orders, Shipping notifications,Subscriptions| *Transactional*|


However, firstly

> **Gmail labels are not the Ground Truth.**

Because I see a lot of promotional emails under Primary Cateory and Updates Category. And newsletters under Promotions and a lot of brand promotions under Updates.

So, clearly it's not very reliable. Hence, I am going to **use Gmail labels as only hints** for my sampling process.

And secondly,

On exploring mail stats and looking at my mails, I think:
- My mailbox is primarily dominated by social media updates which I never look at. So, I'll put these in the "Spam" bucket because I don't want them and they ave filled my inbox.
- Primary Section has a lot of mails that I would want to read like newsletters, important updates, etc. But it has a lot of promotions too. I'll put these under the "Important" bucket for now and we can filter out thhe important ones in the manual labelling part.
- Some of the mails in the Updates section are important to me and I would want to actually read them like newsletters, delivery updates, transactions, etc. So, I'll put these under the "Important" bucket and again, filter them out in the manual labelling part.
- Promotions are present everywere in my mailbox except the Social Category. Promotions tab has some emails that I would want to read but I believe Primary and Updates are better sections to pick important emails from. Hence, I'll put these under the "Promo" Bucket.


### Target Sample Size 

Based on this understanding of Gmail labels, my mailbox constituents and personal relevance of the mails, let's pick a sample of size around 1200 with the following composition:


| Bucket   (Potential)                | Count | Why                         |Gmail Category|
| ------------------------ | ----- | --------------------------- |-------------------------|
| Important| 500   | Not very large in number, important and hence, High precision needed       |Primary / Updates|
| Promotions | 300   | Second most common class and unimportant       |Promotions|
| Spam   | 400   | Most Common class and unimportant |Social|

These buckets are the potential buckets based on the Gmail labels we have. After taking the sample, I'll manually label the mails under these 3 buckets based on my personal relevance, which will be the final labels used for training the ML model.

## Creating Helper Columns for Sampling

# Potential Personal Relevance Flags based on Gmail category flags
df["is_important"] = (
    df["labels"].str.contains("CATEGORY_PRIMARY", regex=False, na=False) |
    df["labels"].str.contains("CATEGORY_UPDATES", regex=False, na=False)
)

df["is_promo"] = df["labels"].str.contains("CATEGORY_PROMOTIONS", regex=False, na=False)
df["is_spam"] = df["labels"].str.contains("CATEGORY_SOCIAL", regex=False, na=False)


# Quick sanity check
df[["is_important", "is_promo", "is_spam"]].mean()


## Important Bucket as almost 32% of the mails, probably because we combined both Primary and Updates section into one bucket.
## Rest of the stats look aligned with my understanding of the mailbox.

### Sample Important emails

primary_df = df[df["is_important"]]

sample_primary = (
    primary_df
    .groupby("sender_domain", group_keys=False)
    .apply(lambda x: x.sample(min(len(x), 3), random_state=42))
)

sample_primary = sample_primary.sample(
    n=min(500, len(sample_primary)),
    random_state=42
)

print("Primary sample:", len(sample_primary))


### Sample Promotional emails

promo_df = df[df["is_promo"]]

sample_promo = (
    promo_df
    .groupby("sender_domain", group_keys=False)
    .apply(lambda x: x.sample(min(len(x), 2), random_state=42))
)

sample_promo = sample_promo.sample(
    n=min(300, len(sample_promo)),
    random_state=42
)

print("Promo sample:", len(sample_promo))


### Sample Spam Emails

spam_df = df[df["is_spam"]]

sample_other = (
    spam_df
    .groupby("sender_domain", group_keys=False)
    .apply(lambda x: x.sample(min(len(x), 2), random_state=42))
)

sample_other = sample_other.sample(
    n=min(400, len(sample_other)),
    random_state=42
)

print("Other sample:", len(sample_other))


spam_df["sender_domain"].nunique()

## There are only 11 sender domains in the Social Category (probably because we have limited popular social media apps I have my account on).
## Hence, we should not limit the mails at 2 per sender domain to obtain the desired sample size.

sample_other = (
    spam_df
    .groupby("sender_domain", group_keys=False)
    .apply(lambda x: x.sample(
        n=max(1, int(400 * len(x) / len(spam_df))),
        random_state=42
    ))
)

print("Spam sample:", len(sample_other))

### Combine all samples

sampled_df = pd.concat(
    [sample_primary, sample_promo, sample_other],
    ignore_index=True
)

print("Total sampled:", len(sampled_df))


## Prepare CSV for manual labeling

#Choosing a subset of required columns
label_df = sampled_df[
    [
        "message_id",
        "subject",
        "snippet",
        "sender",
        "sender_domain",
        "internal_date",
        "is_important",
        "is_promo",
        "is_spam"
    ]
].copy()


#Converting internal date from unix timestamp in milliseconds into date
label_df["internal_date"] = (
    pd.to_datetime(label_df["internal_date"], unit="ms", utc=True)
      .dt.tz_convert("Asia/Kolkata")
      .dt.date
)

#Creating plcaeholder column for manual labels
label_df["label"] = ""   # will fill this manually

#Converting into CSV
label_df.to_csv("../data/emails_to_label.csv", index=False)

#Looking at top 5 rows
label_df.head()


label_df["internal_date"].dtype

## Checking if the samples are evenly spread over time or not
label_df["internal_date"] = pd.to_datetime(label_df["internal_date"])


flags = ["is_important", "is_promo", "is_spam"]

summary = {
    flag: label_df.loc[label_df[flag], "internal_date"].dt.year.value_counts()
    for flag in flags
}

summary_df = pd.DataFrame(summary).fillna(0).astype(int).sort_index()
summary_df


## Looks like the sampling has chosen emails from all the past 9 years. 2017, 2018 and 2019 have lower proportion of emails probably because I started using internet a lot after 2019.

## Criteria for Manual Labelling

| Label         | Meaning             |
| ------------- | ------------------- |
| `important`   | I’d open & act    |
| `promotional` | Legit but ignorable |
| `spam`        | I don’t want this |
